requirements.txt
fastapi
uvicorn
pydantic-settings
openai
langchain
python-dotenv
ruff
numpy
Pillow
ultralytics
pydantic
mediapipe
opencv-python
python-multipart